smart-task-planner/backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.10.3
openai==1.57.3
